Renames base dto
