+++ apps/frontend/src/app/models/base.dto.ts
