+++ apps/frontend/src/app/models/comment.dto.ts
@@ -1,5 +1,5 @@
 import { Model } from '../utils/types';
-import { Dto } from './dto.base';
+import { Dto } from './base.dto';
 import { PlanDto } from './plan.dto';
 import { UserDto } from './user.dto';
