+++ apps/frontend/src/app/models/login-user.dto.ts
@@ -1,5 +1,5 @@
 import { Model } from '../utils/types';
-import { Dto } from './dto.base';
+import { Dto } from './base.dto';
 import { User } from './user.dto';
 
 /**
