+++ apps/frontend/src/app/models/password-update.dto.ts
@@ -1,5 +1,5 @@
 import { Model } from '../utils/types';
-import { Dto } from './dto.base';
+import { Dto } from './base.dto';
 import { UserUpdate } from './user-update.dto';
 
 /**
