+++ apps/frontend/src/app/models/plan.dto.ts
@@ -1,5 +1,5 @@
 import { Model } from '../utils/types';
-import { Dto } from './dto.base';
+import { Dto } from './base.dto';
 import { GameDto } from './game.dto';
 import { UserDto } from './user.dto';
